Fix several bug regarding handling of promises
Fixed and added tests for:
- expectSaga ignored the silenceTimeout flag when any kind of fork was involved.
- forks caused the exportSaga timeout period to be extended. Whenever a fork was added expectSaga would reset the timeout. Adding a fork once would result in timeout period times 2. Keep adding forks and expectSaga would never time out.

Added test for (to make sure nothing broke with this change):
- expectSaga waiting for all promises to resolve (if they resolve within the timeout)

The fix is to move the stopDirty behaviour, so that getAllPromises never resolves before all promises are resolved. This makes sure that cancelMainTask doesn’t have to call scheduleStop, as that call caused issues with silenceTimout not being passed, and the timeout resetting.