Emphasize external documentation in readme
There is some really great documentation on the external Gitbook site but it's not at all obvious it exists when glancing at the readme. The external links are tucked away at the bottom of the page under a small 'Getting Started' section. This new layout is based on how redux and redux-saga has organized their documentation links.